move get_backend to pynwb init
